[BORA00012431] wlan: Fix incompatible args in DBGLOG
[Detail]
This patch only fix coverities with incompatible args in DBGLOG,
which includes:
BORA00012431 for CID 70371
BORA00012792 for CID 4140463
BORA00012788 for CID 4140459
BORA00012768 for CID 4140443
BORA00012661 for CID 361862
BORA00012791 for CID 4140462

MTK-Commit-Id: 4c302b4e1fa81c35139dbc4b464350d010f03e5c

CR-Id: BORA00012431
Feature: Others
Reviewed-on: http://gerrit.mediatek.inc:8080/1274184
CheckPatch: Check Patch <srv_checkpatch@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_mspautosanity@mediatek.com

+++ mgmt/rlm.c
@@ -2314,8 +2314,8 @@ rlmRecOpModeBwForClient(uint8_t ucVhtOpModeChannelWidth, struct BSS_INFO *prBssI
 				"VHT_OP != OPmode:%d, follow VHT_OP to VHT_OP:%d HT_OP:%d\n",
 				ucVhtOpModeChannelWidth,
 				prBssInfo->ucVhtChannelWidth,
-				(prBssInfo->ucHtOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH) >>
-				HT_OP_INFO1_STA_CHNL_WIDTH_OFFSET);
+				(uint8_t) (prBssInfo->ucHtOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH)
+				>> HT_OP_INFO1_STA_CHNL_WIDTH_OFFSET);
 		} else
 #endif
 		{
@@ -2333,7 +2333,7 @@ rlmRecOpModeBwForClient(uint8_t ucVhtOpModeChannelWidth, struct BSS_INFO *prBssI
 				"VHT_OP != OPmode:%d, follow VHT_OP to VHT_OP:%d HT_OP:%d\n",
 				ucVhtOpModeChannelWidth,
 				prBssInfo->ucVhtChannelWidth,
-				(prBssInfo->ucHtOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH) >>
+				(uint8_t) (prBssInfo->ucHtOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH) >>
 				HT_OP_INFO1_STA_CHNL_WIDTH_OFFSET);
 		} else
 #endif
@@ -4610,7 +4610,7 @@ rlmChangeOwnOpInfo(struct ADAPTER *prAdapter, struct BSS_INFO *prBssInfo)
 
 			DBGLOG(RLM, INFO, "Update BSS[%d] HT Channel Width Info to bw=%d sco=%d\n",
 				prBssInfo->ucBssIndex,
-				((prBssInfo->ucHtOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH)
+				(uint8_t) ((prBssInfo->ucHtOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH)
 					>> HT_OP_INFO1_STA_CHNL_WIDTH_OFFSET),
 				prBssInfo->eBssSCO);
 		}
@@ -4866,7 +4866,7 @@ rlmCheckOpChangeParamForClient(struct BSS_INFO *prBssInfo, uint8_t ucChannelWidt
 				"Can't change BSS[%d] OP BW to:%d for PeerOpBw:%d fg40mBwAllowed:%d\n",
 					prBssInfo->ucBssIndex,
 					ucChannelWidth,
-					(prStaRec->ucHtPeerOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH),
+					(uint8_t) (prStaRec->ucHtPeerOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH),
 					prBssInfo->fg40mBwAllowed);
 				return FALSE;
 			}
@@ -4903,7 +4903,7 @@ rlmCheckOpChangeParamForClient(struct BSS_INFO *prBssInfo, uint8_t ucChannelWidt
 						"Can't change BSS[%d] OP BW to:%d for PeerOpBw:%d fg40mBwAllowed:%d\n",
 						prBssInfo->ucBssIndex,
 						ucChannelWidth,
-						(prStaRec->ucHtPeerOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH),
+						(uint8_t) (prStaRec->ucHtPeerOpInfo1 & HT_OP_INFO1_STA_CHNL_WIDTH),
 						prBssInfo->fg40mBwAllowed);
 					return FALSE;
 				}
